# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления кошельками",
        "title": "Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/wallet": {
            "post": {
                "description": "Выполняет операцию пополнения или снятия средств с кошелька",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Применить операцию к кошельку",
                "parameters": [
                    {
                        "description": "Данные операции",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Новый баланс",
                        "schema": {
                            "$ref": "#/definitions/http.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Кошелек не найден",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{id}": {
            "get": {
                "description": "Возвращает текущий баланс указанного кошелька",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Получить баланс кошелька",
                "parameters": [
                    {
                        "type": "string",
                        "example": "11111111-1111-1111-1111-111111111111",
                        "description": "ID кошелька",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Баланс кошелька",
                        "schema": {
                            "$ref": "#/definitions/http.BalanceResponse"
                        }
                    },
                    "404": {
                        "description": "Кошелек не найден",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.ApplyRequest": {
            "type": "object",
            "required": [
                "amount",
                "operationType",
                "walletId"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000
                },
                "operationType": {
                    "type": "string",
                    "example": "DEPOSIT"
                },
                "walletId": {
                    "type": "string",
                    "example": "11111111-1111-1111-1111-111111111111"
                }
            }
        },
        "http.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        }
    }
}